% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_table_avail.R
\name{check_table_avail}
\alias{check_table_avail}
\title{Check if a table is available in a connection and schema}
\usage{
check_table_avail(conn, table)
}
\arguments{
\item{conn}{an object from \code{DBI::dbConnect}.}

\item{table}{character. A table name.}
}
\value{
Logical.
}
\description{
Check if a table is available in a connection and schema
}
