% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_febre_amarela.R
\name{get_febre_amarela}
\alias{get_febre_amarela}
\title{Get febre amarela confirmed cases}
\usage{
get_febre_amarela(
  agg,
  agg_time,
  ano,
  sexo = NULL,
  idade_a = NULL,
  idade_b = NULL,
  psql_user = NULL,
  psql_pwd = NULL
)
}
\arguments{
\item{agg}{character. Spatial aggregation level. \code{uf_res} for UF of residence. \code{regsaude_449_res} for regiao de saude of residence. \code{mun_res} for municipality of residence.}

\item{agg_time}{character. Time aggregation level. \code{year} for yearly data. \code{month} for monthly data. \code{week} for weekly data. Defaults to \code{year}.}

\item{ano}{vector. Year of the case.}

\item{sexo}{character. Sex of the case. \code{Masculino} for males, \code{Feminino} for females and \code{Ignorado} for unknown.}

\item{idade_a}{numeric. Minimum age of the deceased, in years.}

\item{idade_b}{numeric. Maximum age of the deceased, in years.}

\item{psql_user}{character. psql username. If not provided, the function will look for it on renviron.}

\item{psql_pwd}{character. psql password. If not provided, the function will look for it on renviron.}
}
\value{
A tibble.
}
\description{
Get febre amarela confirmed cases
}
