% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_psql_parameters.R
\name{get_psql_parameters}
\alias{get_psql_parameters}
\title{Get psql connection parameters from renviron}
\usage{
get_psql_parameters()
}
\value{
a list.
}
\description{
Get psql connection parameters from renviron
}
