% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dengue.R
\name{get_dengue}
\alias{get_dengue}
\title{Get dengue confirmed cases}
\usage{
get_dengue(agg, agg_time, ano, psql_user = NULL, psql_pwd = NULL)
}
\arguments{
\item{agg}{character. Spatial aggregation level. \code{mun_res} for municipality of residence.}

\item{agg_time}{character. Time aggregation level. \code{year} for yearly data. \code{month} for monthly data. \code{week} for weekly data. Defaults to \code{year}.}

\item{ano}{vector. Year of the case.}

\item{psql_user}{character. psql username. If not provided, the function will look for it on renviron.}

\item{psql_pwd}{character. psql password. If not provided, the function will look for it on renviron.}
}
\value{
A tibble.
}
\description{
Get dengue confirmed cases
}
